% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{assemble_genes}
\alias{assemble_genes}
\title{Assembles a gene list from mutual information output, manual gene selections, and DE genes.}
\usage{
assemble_genes(
  mutualinfopath,
  manualpath,
  gene_list,
  mutual_info_percentile = 0.95
)
}
\arguments{
\item{mutualinfopath}{path to output from the mutual information python function.}

\item{manualpath}{path to a .xlsx file containing the names of manually-selected genes.}

\item{gene_list}{the vector of gene names output from de_combination_genes.}

\item{mutual_info_percentile}{percentile cutoff for mutual information genes. Default 0.95.}
}
\value{
a vector of gene names.
}
\description{
Assembles a gene list from mutual information output, manual gene selections, and DE genes.
}
