% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{chunk_permutations}
\alias{chunk_permutations}
\title{Divides the permutations into a list of chunks.}
\usage{
chunk_permutations(combinations, optimal_chunk_number)
}
\arguments{
\item{combinations}{an n-by-2 integer of permutations between all pairs of clusters}

\item{optimal_chunk_number}{number of chunks to produce}
}
\value{
a list of permutations divided into optimal_chunk_number of chunks
}
\description{
Divides the permutations into a list of chunks.
}
