% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{de_combination_genes}
\alias{de_combination_genes}
\title{Finds the genes with the highest DE score for all pairs of clusters.
If multiple genes are DE between the same pair of clusters, the gene with the highest DE score is retained.}
\usage{
de_combination_genes(de_genes)
}
\arguments{
\item{de_genes}{a data.frame of DE genes, generated using the seurat_de_genes function.}
}
\value{
a vector of gene names.
}
\description{
Finds the genes with the highest DE score for all pairs of clusters.
If multiple genes are DE between the same pair of clusters, the gene with the highest DE score is retained.
}
