% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{find_optimal_chunk_number}
\alias{find_optimal_chunk_number}
\title{Identifies the largest number of chunks the permutations can be divided into.}
\usage{
find_optimal_chunk_number(combinations)
}
\arguments{
\item{combinations}{an n-by-2 integer of permutations between all pairs of clusters}
}
\value{
the number of chunks the permutations should be divided into
}
\description{
Identifies the largest number of chunks the permutations can be divided into.
}
