% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{get_mutual_info_data}
\alias{get_mutual_info_data}
\title{Saves DE gene expression data from a Seurat object as a .csv file.
This file can be used as input to python code to calculate mutual information.}
\usage{
get_mutual_info_data(seuratobject, de_genes)
}
\arguments{
\item{seuratobject}{an SCTransformed, integrated Seurat object.}

\item{de_genes}{a data.frame of DE genes, generated using the seurat_de_genes function.}
}
\value{
a path to the .csv file containing the SCT-corrected counts for all unique DE genes in the input Seurat object.
}
\description{
Saves DE gene expression data from a Seurat object as a .csv file.
This file can be used as input to python code to calculate mutual information.
}
