% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{identify.de.genes.wilcoxon}
\alias{identify.de.genes.wilcoxon}
\title{Identifies DE genes using a Wilcoxon rank-sum test.}
\usage{
\method{identify}{de.genes.wilcoxon}(combinations, seuratobject, log_threshold, min_pct)
}
\arguments{
\item{combinations}{a Matrix of identity combinations}

\item{seuratobject}{a Seurat object}

\item{log_threshold}{the minimum log2FC value}

\item{min_pct}{minimum expression value}
}
\value{
a data.frame of the 20 genes with the highest DE score for each combination of identities
}
\description{
Identifies DE genes using a Wilcoxon rank-sum test.
}
