% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid.R
\name{seurat_de_genes}
\alias{seurat_de_genes}
\title{Finds DE genes between all pairs of clusters in a Seurat object.
Calculates a DE score equal to -log10(padj). Max score = 300.}
\usage{
seurat_de_genes(seuratobject, log_threshold = 2, min_pct = 0.5)
}
\arguments{
\item{seuratobject}{a Seurat object with cluster identities in the Idents slot.}

\item{log_threshold}{the minimum log2FC for a gene to be considered DE.}

\item{min_pct}{the minimum fraction of gene expression in one cluster for a gene to be considered DE.}
}
\value{
a data.frame of differentially-expressed genes between all pairs of clusters.
}
\description{
Finds DE genes between all pairs of clusters in a Seurat object.
Calculates a DE score equal to -log10(padj). Max score = 300.
}
